Web site map
Commit "sitemap.php", a first prototype that generates an HTML site
map for the REDUCE web site in the form of a table of URLs with page
titles (where they exist) and last modified dates.  This is intended
for human users, not web crawlers!  The current version is not (yet)
linked from the web site and not very pretty.  It needs more work, and
in particular CSS and JavaScript support to collapse sub-directories.

I have also tweaked "index.php" so that "sitemap.php" reports the
correct page title, because at present "index.php" doesn't follow the
pattern of other PHP pages.