Web site map
Add JavaScript (mostly jQuery) and CSS to collapse and expand
sub-folders.  To facilitate this, in the PHP code I annotate each
table row with an attribute of the form data-level=n to show the
nesting level of folder entries, and I pass a level argument through
the recursive function call that processes each sub-folder.  Ensure
that all entries, including HTML files with an empty title element,
have either a page title or a description.

Switch to using Emacs web-mode, which seems to handle the mix of HTML,
CSS, PHP and JavaScript within one file reasonably well (which
php-mode doesn't!), but it indents differently leading to some
inconsequential white space changes.