Web site
This is the first step of a revision of the web site to stop using the
Smarty template engine and just use "raw" PHP.  I have decided that a
template engine is unnecessary for this web site and causes as many
problems as it solves.  In particular, Smarty tries to keep code and
content separate, which is not particularly helpful when some of the
content is generated by code and when the same people write the code
and the content.  Integrating the new sitemap code into the template
system would have been somewhat convoluted, which was mainly what
provoked this change of approach.  So far, I have re-implemented and
checked all the top-level pages, which I think includes all the
difficult ones, but I want to revise the documentation page further to
use PHP rather than JavaScript to generate the list of links to the
user contributed packages in the manual.