REDUCE IDE version 1.9
Improved support for, in particular, indentation.

Trying to improve support for comment statements suggests that I need
better parsing, so as the next major development I plan to provide a
new parsing facility that will parse from the beginning of the buffer
(rather than locally) and initially just handle comment statements.
This will be analogous to the parsing facility provided by Emacs that
uses the syntax table.  This works well for REDUCE % and /**/ comments
(and strings), but not for comment statements, which are particularly
awkward to parse backwards reliably!  (When parsing locally, I need
to check that keywords and terminators are not within comments or
strings, which gets messy!)