Substantial additions and corrections in this commit. New module efreim.red: Implement complex split functions for the Jacobi elliptic functions for the cases when the modulus is either real or purely imaginary.
efjacobi.red:  The addition rules and double 'angle' rules for the Jacobi
elliptic functions have been moved to a new rulelist 'JacobiAdditionRules'
which is NOT active by default. 

Also the rhs of the rules for the 9 Glaisher functions have been changed to
a more 'democratic' form although the previous forms may be recovered by
activating the rulelist no_glaisher.

There are substantial changes to the numerical evaluation code particularly
for the Jacobi functions and elliptic integrals. Firstly the code for the 12
Jacobi functions is now via theta functions as this method is valid for all
complex values of both the argument and the modulus.  The previous AGM-based
method was only really reliable when the modulus was real and <1. Also
there was a problem with the code for jacobidn leading to erroneous results
in many cases even when the modulus was real and <1.  This dated back to the
original 1994 implementation. This had knock on effects on the evaluation of
nd,sd,ds, cd and dc.

The evaluation of complete elliptic integrals now uses Carlson's method of
symmetric elliptic integrals and now works for all complex values of the
modulus.

As a consequence of these changes it is now necessary for both
of the switches ROUNDED and COMPLEX to be ON even if all the arguments are
numerical and real.

The loop termiation condition used in the numerical evaluation of theta
functions and their derivatives has been improved. Previously in certain cases
with a complex argument the loop could terminate prematurely leading to
substantial rounding errors.

Minor corrections and deletions from other modules of the package.

ellipfn.tst & ellipfn.rlg: Some additional examples of the numerical evaluation
of the Jacobi functions, their inverses and the complete elliptic integrals
together with a number of consistency tests.

ellipfn.tex & manual.pdf: a few paragraphs added to document the corrections
and new features in this commit.