ellipfn.tex Correct a typo and add a caveat on the use of eighth period rules.
packages/ellipfn/efjacobi.red:
Add eighth period rules for all 12 Jacobi functions. These rules require the
switch PRECISE_COMPLEX to be ON both when the rules are defined and invoked to
prevent incorrect simplification of complex square roots etc..
The rule for dn((K(k)+i*K'(k))/2, k) differs from that on the DLMF:NIST site
which I believe is incorrect (possibly due to the assumption that K'(k) = K(k')
for all complex values of k -- see the paragraph below).

packages/ellipfn/efellint.red:
The numerical evaluation rules for the complete elliptic integrals of the first
and second kinds have been extended (& corrected in places). They now return
the correct principal value for all complex arguments except, of course, at
singular points. 

Formerly the evaluation of the  complementary integrals K'(k)
and E'(k) used the identities K'(k) = K(k'); E'(k) = E(k') for all complex
values of k. These identities are not valid in the second and third quadrants
of the complex plane and so were corrected. As a consequence the two
complementary integrals are no long even functions of the modulus and the
erroneous rules have been deleted. Maple 16 gets all this wrong!

The numerical evaluation of Legendre's form of the incomplete integral of the
1st kind has been improved and now works for all values of the modulus k
and for all arguments phi where -pi/2 <= repart(phi) <= pi/2.

packages/ellipfn/efjacinv.red:
Minor reformatting.

packages/ellipfn.rlg:
Changes due to the correction of the algorithm for ellipticE and to different
but insignificant rounding errors.